refactor(intern): align task management imports with barrel and remove unused import
